Added signup component
