--- Client - React/v1/.env.example
@@ -1,1 +1,0 @@
-REACT_APP_SERVER_URL=https://eatmyurl.ml
