--- Client - React/v1/src/index.css
@@ -1,74 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins&display=swap");
-* {
-  margin: 0%;
-  padding: 0%;
-  font-family: "Poppins", sans-serif;
-  overflow: hidden;
-}
-
-body {
-  height: 100vh;
-  width: 100vw;
-  background: #fcedea;
-}
-.card {
-  height: 50vh;
-  background: #f4b6b0;
-}
-.center {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-h4 {
-  font-weight: bold;
-  padding: 10px;
-  width: fit-content;
-  height: fit-content;
-  background: #e79392;
-  color: #f4b6b0;
-  margin: 2%;
-}
-.custom-flip-button {
-  background: #e79392;
-  color: #f4b6b0;
-  border: none;
-  border-radius: 5px;
-}
-.custom-button {
-  background: #ef9a9a;
-  font-size: large;
-  border: none;
-  margin-top: 5%;
-  padding: 10px;
-  width: max-content;
-  align-items: center;
-}
-.custom-button:hover {
-  background: #df7c73;
-}
-.custom-button:active {
-  background: #e79392;
-}
-.custom-button:focus {
-  background: #e79392;
-}
-.button-bottom {
-  margin: 10px;
-  position: absolute;
-  bottom: 0;
-  width: max-content;
-  right: 0;
-}
-.display-none {
-  display: display;
-}
-p {
-  text-shadow: 5px 5px 5px #000000;
-}
-.footer-text {
-  margin-left: 2px;
-  color: royalblue;
-  font-weight: bold;
-  cursor: pointer;
-}
