+++ ClientReact/v1/.env
@@ -1,0 +1,2 @@
+REACT_APP_SERVER_URL=https://eatmyurl.ml
+SKIP_PREFLIGHT_CHECK=true
