+++ ClientReact/v1/package.json
@@ -7,10 +7,12 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.4",
+    "bootstrap": "^5.1.3",
     "react": "^17.0.2",
     "react-card-flip": "^1.1.3",
     "react-dom": "^17.0.2",
     "react-reveal": "^1.2.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "react-share": "^4.4.0",
     "react-toastify": "^8.0.3",
