+++ ClientReact/v1/public/manifest.json
