+++ ClientReact/v1/src/context/textContext.js
