+++ ClientReact/v1/src/index.css
@@ -1,0 +1,153 @@
+@import url("https://fonts.googleapis.com/css2?family=Poppins&display=swap");
+* {
+  margin: 0%;
+  padding: 0%;
+  font-family: "Poppins", sans-serif;
+  overflow: hidden;
+}
+
+body {
+  height: 100vh;
+  overflow-y: scroll;
+  width: 100vw;
+  background: #fcedea;
+}
+.card {
+  height: 50vh !important;
+  background: #f4b6b0 !important;
+}
+.center {
+  display: flex ;
+  justify-content: center ;
+  align-items: center ;
+}
+h4 {
+  font-weight: bold;
+  padding: 10px;
+  width: fit-content;
+  height: fit-content;
+  background: #e79392;
+  color: #f4b6b0;
+  margin: 2%;
+}
+.custom-flip-button {
+  background: #e79392;
+  color: #f4b6b0;
+  border: none;
+  border-radius: 5px;
+}
+.custom-button {
+  background: #ef9a9a;
+  font-size: large;
+  border: none;
+  margin-top: 5%;
+  padding: 10px;
+  width: max-content;
+  align-items: center;
+}
+.custom-button:hover {
+  background: #df7c73;
+}
+.custom-button:active {
+  background: #e79392;
+}
+.custom-button:focus {
+  background: #e79392;
+}
+.button-bottom {
+  margin: 10px;
+  position: absolute;
+  bottom: 0;
+  width: max-content;
+  right: 0;
+}
+.display-none {
+  display: display;
+}
+p {
+  text-shadow: 5px 5px 5px #000000;
+}
+.footer-text {
+  margin-left: 2px;
+  color: royalblue;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+/*CSS for toggling signup page*/
+
+
+.hide{
+  display: none;
+}
+
+.btn{
+  padding: 0.3rem;
+  margin: 0.3rem;
+  border-radius: 0.5rem;
+  border: 2px solid black;
+}
+
+/*CSS for main input container*/
+.signup{
+  border: 2px solid #df7c73;
+  border-radius: 0.5rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 0.5rem;
+  padding: 0.5rem;
+  background-color: #fcedea;
+}
+.bg{
+  background-color: #ef9a9a;
+}
+/*CSS for input section in container*/
+.inputsection{
+  border: 2px solid #df7c73;
+  border-radius: 0.5rem;
+  padding-left: 3rem;
+  padding-right: 3rem;
+  background-color: #f4b6b0;
+ margin-bottom: 2rem;
+  /* padding: 5rem; */
+}
+
+/*CSS separating each input field*/
+.inputsection .inputwrap{
+  padding: 0.8rem;
+}
+
+.inputwrap input{
+  border-radius: 0.6rem;
+  height: 2.5rem;
+  padding: 0.6rem;
+}
+
+/*Signup header*/
+.header{
+  font-size: 3rem;
+  padding: 0.2rem;
+  font-weight: bolder;
+  color: #733030;
+}
+/*Button container*/
+.btnwrap{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+/*Button CSS*/
+.btn{
+  border: 2px solid #733030 !important;
+  background-color: #df7c73 !important;
+  border-radius: 0.6rem !important;
+  padding: 0.2rem;
+  color: #fff !important;
+  cursor: pointer;
+}
+
+.btn:hover{
+  background-color: #e79392 !important;
+}
+
