+++ ClientReact/v2/public/manifest.json
